# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: NOVA API
  description: Voice and Visual Companion platform REST API — Admin Console, Audit, and Observability
  version: 0.1.0
  contact:
    name: Leadup Technologies
  license:
    name: Proprietary
servers:
  - url: http://localhost:3004
    description: Admin service (local development)
paths:
...
  /admin/organizations:
    get:
      tags:
        - Admin Console
      summary: List organizations for the current user
      operationId: listOrganizations
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Organizations list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedOrganizations'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/organizations/{id}:
    get:
      tags:
        - Admin Console
      summary: Get a single organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Organization details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Admin Console
      summary: Update an organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrganization'
      responses:
        '200':
          description: Updated organization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /admin/feature-flags:
    get:
      tags:
        - Feature Flags
      summary: List all feature flags
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Feature flags list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FeatureFlag'

    patch:
      tags:
        - Feature Flags
      summary: Update a feature flag
      security:
        - bearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFeatureFlag'
      responses:
        '200':
          description: Updated feature flag
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureFlag'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /admin/incidents:
    get:
      tags:
        - Observability
      summary: List incident events
      security:
        - bearerAuth: []
      parameters:
        - name: severity
          in: query
          schema:
            type: string
            enum: [info, warning, error, critical]
        - name: category
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated incidents

    post:
      tags:
        - Observability
      summary: Create an incident event
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIncident'
      responses:
        '201':
          description: Incident created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentEvent'
...
  /admin/incidents/{id}/resolve:
    patch:
      tags:
        - Observability
      summary: Resolve an incident
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Incident resolved
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        plan:
          type: string
        created_at:
          type: string
          format: date-time
    PaginatedOrganizations:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Organization'
        page:
          type: integer
        pageSize:
          type: integer
        totalItems:
          type: integer
        totalPages:
          type: integer
    UpdateOrganization:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        plan:
          type: string
          enum: [free, pro, enterprise]
...
    FeatureFlag:
      type: object
      properties:
        id:
          type: string
          format: uuid
        key:
          type: string
        description:
          type: string
        enabled:
          type: boolean
        rollout_percentage:
          type: integer
        allowed_org_ids:
          type: array
          items:
            type: string
            format: uuid
        allowed_user_ids:
          type: array
          items:
            type: string
            format: uuid
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UpdateFeatureFlag:
      type: object
      properties:
        enabled:
          type: boolean
        rollout_percentage:
          type: integer
          minimum: 0
          maximum: 100
        allowed_org_ids:
          type: array
          items:
            type: string
            format: uuid
        allowed_user_ids:
          type: array
          items:
            type: string
            format: uuid
        metadata:
          type: object
          additionalProperties: true
...
    CreateIncident:
      type: object
      required:
        - severity
        - category
        - title
      properties:
        severity:
          type: string
          enum: [info, warning, error, critical]
        category:
          type: string
          maxLength: 50
        title:
          type: string
          maxLength: 255
        description:
          type: string
        source:
          type: string
          maxLength: 100
        metadata:
          type: object
          additionalProperties: true
    IncidentEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        severity:
          type: string
        category:
          type: string
        title:
          type: string
        description:
          type: string
        source:
          type: string
        metadata:
          type: object
          additionalProperties: true
        acknowledged_at:
          type: string
          format: date-time
        acknowledged_by:
          type: string
          format: uuid
        resolved_at:
          type: string
          format: date-time
        resolved_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            type: object
            properties:
              type:
                type: string
                example: https://api.nova.leadup.in/problems/unauthorized
              title:
                type: string
                example: Unauthorized
              status:
                type: integer
                example: 401
              detail:
                type: string
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            type: object
            properties:
              type:
                type: string
                example: https://api.nova.leadup.in/problems/forbidden
              title:
                type: string
                example: Forbidden
              status:
                type: integer
                example: 403
              detail:
                type: string
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              type:
                type: string
                example: about:blank
              title:
                type: string
                example: Not Found
              status:
                type: integer
                example: 404
              detail:
                type: string